requirements.txt
rich==13.7.1
pyyaml==6.0.1
httpx==0.27.0
openai==1.14.0
